requirements.txt
requests
yamllint
ansible-lint
ansible-core
